AI: Refactored the agent logic
Theoretically, there are many agents that can implement a task differently.

In an extreme example, a crane/elevator might have a lorry/container that can transport items. If this were an AI agent, it could get an item and bring it somewhere with the same logic that the clonk does - only how it does that is different.

So, the task tells the AI what to do, the agent tells the task how to do it.